Extracted common elements from ApiUser and DataHubUpload.
