+++ backend/src/main/java/gov/cdc/usds/simplereport/db/model/ApiUser.java
@@ -1,15 +1,10 @@
 package gov.cdc.usds.simplereport.db.model;
 
 import java.util.Date;
-import java.util.UUID;
 
 import javax.persistence.Column;
 import javax.persistence.Embedded;
 import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
-import javax.persistence.Temporal;
-import javax.persistence.TemporalType;
 
 import org.hibernate.annotations.CreationTimestamp;
 import org.hibernate.annotations.DynamicUpdate;
@@ -27,20 +22,8 @@ import gov.cdc.usds.simplereport.db.model.auxiliary.PersonName;
  */
 @Entity
 @DynamicUpdate
-public class ApiUser {
+public class ApiUser extends SystemManagedEntity {
 
-	@Column(updatable = false, nullable = false)
-	@Id
-	@GeneratedValue(generator = "UUID4")
-	private UUID internalId;
-	@Column(updatable = false)
-	@Temporal(TemporalType.TIMESTAMP)
-	@CreationTimestamp
-	private Date createdAt;
-	@Column
-	@Temporal(TemporalType.TIMESTAMP)
-	@UpdateTimestamp
-	private Date updatedAt;
 	@Column(nullable = false, updatable = true, unique = true)
 	@NaturalId(mutable = true)
 	private String loginEmail;
@@ -55,18 +38,6 @@ public class ApiUser {
 		loginEmail = email;
 		nameInfo = name;
 		lastSeen = null;
-	}
-
-	public UUID getInternalId() {
-		return internalId;
-	}
-
-	public Date getCreatedAt() {
-		return createdAt;
-	}
-
-	public Date getUpdatedAt() {
-		return updatedAt;
 	}
 
 	public String getLoginEmail() {
