+++ backend/src/main/java/gov/cdc/usds/simplereport/db/model/DataHubUpload.java
@@ -1,26 +1,21 @@
 package gov.cdc.usds.simplereport.db.model;
 
 import java.util.Date;
-import java.util.UUID;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.EnumType;
 import javax.persistence.Enumerated;
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
 
 import gov.cdc.usds.simplereport.db.model.auxiliary.DataHubUploadStatus;
 
-import org.hibernate.annotations.CreationTimestamp;
 import org.hibernate.annotations.DynamicUpdate;
 import org.hibernate.annotations.Type;
-import org.hibernate.annotations.UpdateTimestamp;
 
 
 @Entity
 @DynamicUpdate
-public class DataHubUpload {
+public class DataHubUpload extends SystemManagedEntity {
 
     // We have to pass config into the constructor because Configuration doesn't work in constructors
     public DataHubUpload() {
@@ -30,24 +25,11 @@ public class DataHubUpload {
         errorMessage = "";
     }
 
-    @Column(updatable = false, nullable = false)
-    @Id
-    @GeneratedValue(generator = "UUID4")
-    private UUID internalId;
-
     // set to "SUCCESS" when done.
     @Column(nullable = false)
     @Type(type = "pg_enum")
     @Enumerated(EnumType.STRING)
     private DataHubUploadStatus jobStatus;
-
-    @Column(updatable = false)
-    @CreationTimestamp
-    private Date createdAt;
-
-    @Column
-    @UpdateTimestamp
-    private Date updatedAt;
 
     @Column
     private int recordsProcessed;
@@ -65,10 +47,6 @@ public class DataHubUpload {
     @Type(type = "jsonb")
     private String responseData;
 
-    public UUID getInternalId() {
-        return internalId;
-    }
-
     public DataHubUploadStatus getJobStatus() {
         return jobStatus;
     }
@@ -76,14 +54,6 @@ public class DataHubUpload {
     public DataHubUpload setJobStatus(DataHubUploadStatus jobStatus) {
         this.jobStatus = jobStatus;
         return this;
-    }
-
-    public Date getCreatedAt() {
-        return createdAt;
-    }
-
-    public Date getUpdatedAt() {
-        return updatedAt;
     }
 
     public int getRecordsProcessed() {
